requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
tenacity>=8.2.0